finished
Tic-Tac-Toe Game

Overview
Welcome to the Tic-Tac-Toe game repository! This simple yet classic game is a perfect demonstration of your programming skills. Tic-Tac-Toe, also known as Noughts and Crosses, is a two-player game where the objective is to be the first to get three of your marks (either "X" or "O") in a row, column, or diagonal on a 3x3 grid.

This GitHub repository contains a HTML, CSS and JavaScript implementation of the Tic-Tac-Toe game. It's an ideal project for beginners looking to learn the basics of JavaScript or for more experienced developers wanting to brush up on their skills. The code is clean, well-documented, and easy to understand.

Features
Two-Player Mode: Play against a friend locally, taking turns to make your moves.

Interactive Board: A user-friendly interface that displays the current state of the game grid.

Win Detection: The game automatically checks for a win or a draw and announces the winner or declares a draw.